fix nox tests, move functions from merge.py do merge_functions.py as merge.py god more than 1000 lines of code
